Clarification for MPL-2.0 license, include required Exhibit A
The dynasm and dynasmrt crates claim an MPL-2.0 license in the
Cargo.toml and by including the LICENSE file, but the text of the
license requires a specific declaration called the "source code form" to
be made. Without this declaration, it's not possible to strictly comply
with the license. This patch adds Exhibit A to the Cargo.toml where it
should be unobtrusive but sufficient for license adherence.